moved isGeometryValid into lib
